+++ CHANGELOG.md
@@ -8,6 +8,12 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 * The MINOR component is used whenever a version has backwards-compatible profile changes. This also indicates that the user can not switch back to a previous MINOR version without using a backup.
 * The PATCH component is used for versions that do not contain profile format changes. Users can freely switch between PATCH versions without risk of data loss.
 
+## [Unreleased]
+
+### Fixed
+
+* Fix error when unequipping items that are bound via slot IDs [#191]
+
 ## [1.14.4] - 2024-02-11
 
 ### Added
@@ -1379,6 +1385,7 @@ The format of this changelog is based on [Keep a Changelog](https://keepachangel
 [0.4.0]: https://github.com/Snakybo/Clicked/releases/tag/0.4.0
 [0.3.0]: https://github.com/Snakybo/Clicked/releases/tag/0.3.0
 
+[#191]: https://github.com/Snakybo/Clicked/issues/191
 [#189]: https://github.com/Snakybo/Clicked/issues/189
 [#186]: https://github.com/Snakybo/Clicked/issues/186
 [#184]: https://github.com/Snakybo/Clicked/issues/184